Clean up appState
