+++ frontend/src/UI/hooks/useNavItems.tsx
@@ -1,8 +1,11 @@
 import { useAppState } from "../../business/overmind";
+import {
+  privateNavItems,
+  publicNavItems,
+} from "../../configuration/navigation";
 
 export function useNavItems() {
   const { isSignedIn } = useAppState().auth;
-  const { publicNavItems, privateNavItems } = useAppState().app;
 
   if (isSignedIn) {
     return [...privateNavItems, ...publicNavItems];
