+++ frontend/src/business/overmind/Auth/actions.ts
@@ -1,7 +1,7 @@
 /* eslint-disable no-param-reassign */
 import { AxiosError } from "axios";
 import { Context } from "..";
-import { Page } from "../../models";
+import { Page } from "../../../configuration/navigation";
 
 export const showSignInPage = ({ state }: Context) => {
   state.app.currentPage = Page.SIGNIN;
