+++ frontend/src/configuration/chart.ts
@@ -1,0 +1,7 @@
+export const COLORS = [
+  "#ef476fff",
+  "#ffd166ff",
+  "#06d6a0ff",
+  "#118ab2ff",
+  "#073b4cff",
+];
